# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger documentation for the BitSong Media Server service API.",
        "title": "bitsongms API Docs",
        "contact": {
            "name": "BitSong",
            "email": "hello@bitsong.io"
        },
        "license": {
            "name": "CC0",
            "url": "https://creativecommons.org/share-your-work/public-domain/cc0/"
        },
        "version": "0.1"
    },
    "host": "localhost:1347",
    "basePath": "/api/v1",
    "paths": {
        "/transcode/{id}": {
            "get": {
                "description": "Get transcode status by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transcode"
                ],
                "summary": "Get transcode status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transcoder"
                        }
                    },
                    "400": {
                        "description": "Failure to parse the id",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorJson"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorJson"
                        }
                    }
                }
            }
        },
        "/upload/audio": {
            "post": {
                "description": "Upload, transcode and publish to ipfs an audio",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Upload and transcode audio file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Audio file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.UploadAudioResp"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorJson"
                        }
                    }
                }
            }
        },
        "/upload/image": {
            "post": {
                "description": "Upload, create and publish to ipfs an image",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Upload and create image file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.UploadImageResp"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorJson"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Transcoder": {
            "type": "object",
            "properties": {
                "db": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "server.ErrorJson": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/server.ErrorJsonBody"
                }
            }
        },
        "server.ErrorJsonBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "server.UploadAudioResp": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "track_id": {
                    "type": "string"
                },
                "transcoder_id": {
                    "type": "string"
                }
            }
        },
        "server.UploadImageResp": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                }
            }
        }
    }
}